refactor: firebase.tsからauthorizationの機能だけ別ファイルに切り出し
