+++ __tests__/index.test.tsx
@@ -1,11 +1,11 @@
 import IndexPage from '../src/pages/index';
 import { AuthProvider } from '../src/providers/AuthProvider';
 import { mockServer } from '../src/services/api/mocks/server';
-import { getToken, getUser } from '../src/services/firebase/firebase';
+import { getToken, getUser } from '../src/services/firebase/authentication';
 import { render, screen, waitFor } from '@testing-library/react';
 
 // モック
-jest.mock('../src/services/firebase/firebase');
+jest.mock('../src/services/firebase/authentication');
 const mockGetToken = getToken as jest.MockedFunction<typeof getToken>;
 mockGetToken.mockResolvedValue('testtoken');
 const mockGetUser = getUser as jest.MockedFunction<typeof getUser>;
